Remove associatedtype requirements from TensorOperation.
Having an associatedtype on TensorOperation requires `makeOp` to be
defined with opaque return types as follows:

  `ExecutionContext.makeOp() -> some TFTensorOperation`

However, opaque return types are not available everywhere
yet. Removing the associatedtype requirements for the time-being so
that `makeOp` can be defined as follows:
  `ExecutionContext.makeOp() -> TFTensorOperation`

This also merges the declaration of protocol `TensorOperation` into
`TFTensorOperation.swift`.